refactor: credential issuer body
